Modify Cgpa.c
In this commit, the user interactive texts were modified to make it more
easier to understand. Also the conditional statements were adjusted to
make user inputs to be strictly restricted to a defined range. These
modifications became neccessary due to a user experience feedback and
was implemented to make the program more efficient and easy to use.