Add Input Validator Functions in cgpa.h
In this commit, two functions were added to validate the inputs entered
by the user. This became neccessary as it was noticed following how the
program works prior to this commit that when a user enters a string
literal after the integer or floating point value requested by the
program as the case may be, undefined behaviour sets in. This
consequently leads to the crashing of the program.

With the use and implementation and use of these input validator
functions, the probabilty of the program crashing has been greatly
reduced to the bearest minimum.

Here are the two input validator functions that were added;
- nYrsInputValidator()
- gpInputValidator()